Added show/hide to GoLive Toggle
@blackjid from our recent conversation about show/hide. I need a state
for the GoLive toggle to show when the user is 'live'. I put
ng-show/hide with sample classes of user.onAir and user.offAir I don't
know what it would be, but I think that follows the way it would be
done. Can you add these states to app.js? [#59595110]