hookup metadata
User's 'name' shows on basic action bar.
I think geolocation and tags are placed correct in basic video module.
Added a default to geolocation (anonymous).

@emilioeduardob I added a checkbox to the Go Live form to auto-enable
geolocation to a stream. It is set to checked. If user unchecks, then
the geolocation would be turned off. Can you hook this up? Thanks!
[#59595110]