added flip check to modal
[#62917330] added camera flip checkbox to Go Live modal. Will need to
be hooked up so that it only shows up when there is more than one
camera and that it can be checked on/off. @blackjid NOTE: I can make it
show/hide but it would be hidden and no one could see it yet. So let's
set up the show/hide when it works?