Styled Channel Results
[#63016306] Channel results have been styled. A search term will return
channels that match. However, content needs to be pulled in to display:
They are set up to show the thumbnail of the '0' stream from the
channel stream array, to display the channel name, and to display the
number of streams within the channel array. These will need to be
hooked up. I made my best guess as to how to put it together.
@emilioeduardob @blackjid  See Github issue for details.