Fixes to GoLive Form
[#62917330] Fixed the 'add caption' input. Fixed the inputs that were
too large for mobile. VERY IMPORTANT: The 'form' tag was removed from
the Go Live Modal. It is very, very important that my HTML is not
changed in such a drastic way. This affected all the styles for the
form. Labels, Inputs, Selects, Submits, etc MUST reside within Forms.
That is the correct semantic usage as well as important for styling.
I'm assuming that it was removed by accident, but please watch out for
these kind of tags. Thanks! :-) @emilioeduardob @blackjid